requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Background Tasks
celery>=5.3.0
redis>=4.6.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# PDF Generation
reportlab>=4.0.0
PyPDF2>=3.0.0

# Utilities
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0
requests>=2.31.0